requirements.txt
gradio
opencv-python
Pillow
numpy
deepface